Add files via upload
This is the first BETA Test Release for the Compliance Assessment tool. It is fully functional in "Compliance" mode. For "Maturity" mode the assessment questions from the EASA guidance material "Part-IS oversight approach" are actually inserted only for IS.I/D.OR.240. The questions can be edited individually in ths version.